AuthConfig refactoring
